requirements.txt
Biopython>=1.79
nltk>=3.6.3
rich>=10.12.0
pandas>=1.3.3
